Update Noah compile script to bash.

+++ trunk/NDHMS/compile_offline_Noah.sh
@@ -1,6 +1,11 @@
-#!/bin/sh -f
+#!/bin/bash
 
-./setEnvar.csh
+source ./setEnvar.sh
+
+if [[ "$WRF_HYDRO" -ne 1 ]]; then
+    echo "Please set WRF_HYDRO to be 1 from setEnvar.sh"
+    exit
+fi
 
 rm -f LandModel LandModel_cpl
 cp arc/Makefile.Noah Makefile
@@ -9,12 +14,32 @@ ln -sf Land_models/Noah LandModel
 cat macros LandModel/user_build_options.bak  > LandModel/user_build_options
 make clean ; rm -f Run/wrf_hydro_Noah.exe ; rm -f Run/*TBL ; rm -f Run/*namelist*
 
-
 cat macros LandModel/user_build_options.bak > LandModel/user_build_options
+
 make
+
+
+if [[ $? -eq 0 ]]; then
+    echo
+    echo '*****************************************************************'
+    echo "Make was successful"
+else 
+    echo
+    echo '*****************************************************************'
+    echo "Make NOT successful"
+    exit 1
+fi
+
 cd Run
 mv wrf_hydro.exe wrf_hydro_Noah.exe ; ln -sf wrf_hydro_Noah.exe wrf_hydro.exe
 cp ../Land_models/Noah/Run/*TBL .
 cp ../template/Noah/namelist.hrldas .
 cp ../template/HYDRO/HYDRO.TBL .
 cp ../template/HYDRO/hydro.namelist .
+
+echo
+echo '*****************************************************************'
+echo "The environment variables used in the compile:"
+printenv | egrep -i "(HYDRO|NUDG|PRECIP|CHAN_CONN|^NETCDF|REALTIME|SOIL|WRFIO)" | egrep -v PWD
+
+exit 0
